edit simple components with abstract class
