+++ src/view/list-view.js
@@ -1,26 +1,13 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createListElement = () => (
   `<ul class="trip-events__list">
   </ul`
 );
 
-export default class ListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class ListView extends AbstractView {
 
   get template() {
     return createListElement();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
